cambio de color de cada marciano si dos balas tocan a la vez las barreras
