+++ app/src/main/java/procesos/grp7/spaceinvadersprocesossoftware/BulletCollisionDetector.java
@@ -1,8 +1,11 @@
 package procesos.grp7.spaceinvadersprocesossoftware;
 
+import android.graphics.Color;
+import android.graphics.drawable.BitmapDrawable;
 import android.widget.ImageView;
 
 import java.util.List;
+import java.util.Random;
 
 public class BulletCollisionDetector extends CollisionDetector implements Runnable {
 
@@ -11,6 +14,7 @@ public class BulletCollisionDetector extends CollisionDetector implements Runnab
     private GameActivity activity;
     private boolean fromMarciano;
     private List<ImageView> listaMarcianos;
+    long tiempodecolisionanterior=0;
 
     public BulletCollisionDetector(Bullet bullet, List<ImageView> gameViews, GameActivity activity, boolean fromMarciano, List<ImageView> listaMarcianos) {
         this.bullet = bullet;
@@ -19,15 +23,43 @@ public class BulletCollisionDetector extends CollisionDetector implements Runnab
         this.fromMarciano = fromMarciano;
         this.listaMarcianos = listaMarcianos;
     }
-
+    public void cambiarColor(){
+        Random rnd = new Random();
+        int color = Color.argb(255, rnd.nextInt(256), rnd.nextInt(256), rnd.nextInt(256));
+        for (ImageView m:this.listaMarcianos) {
+            m.setColorFilter(color);
+        }
+    }
+    public void cambiarColorAleatorio(){
+        Random rnd = new Random();
+        int color ;
+        for (ImageView m:this.listaMarcianos) {
+            color=Color.argb(255, rnd.nextInt(256), rnd.nextInt(256), rnd.nextInt(256));
+            m.setColorFilter(color);
+        }
+    }
+    public void cambiarColorComprobandoTiempo(ImageView collider){
+        if (((BitmapDrawable) collider.getDrawable()).getBitmap().sameAs(((BitmapDrawable) activity.getResources().getDrawable(R.drawable.defensas)).getBitmap())){
+            long tiempodecolision=System.currentTimeMillis();
+            if (((tiempodecolisionanterior+100))>=tiempodecolision){
+                tiempodecolisionanterior=tiempodecolision;
+                cambiarColorAleatorio();
+            }else{
+                tiempodecolisionanterior=tiempodecolision;
+                cambiarColor();
+            }
+        }
+    }
     @Override
     public void run() {
         long aliveTime = 0;
         long startTime = System.currentTimeMillis();
+        tiempodecolisionanterior=startTime;
         long actualTime;
         while (aliveTime < Bullet.DURATION) {
             final ImageView collider = detectCollision(gameViews, bullet.getBulletView());
             if (collider != null) {
+                cambiarColorComprobandoTiempo(collider);
                 if (fromMarciano) {
                     if (!listaMarcianos.contains(collider)) {
                         activity.kill(bullet, collider);
