+++ app/src/main/java/procesos/grp7/spaceinvadersprocesossoftware/GameActivity.java
@@ -8,6 +8,7 @@ import android.os.AsyncTask;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
+import android.util.Log;
 import android.view.Display;
 import android.view.MotionEvent;
 import android.view.View;
@@ -88,7 +89,6 @@ public class GameActivity extends PlayActivity implements View.OnTouchListener {
         Thread shipCollisionDetector = new Thread(new ShipCollisionDetector(this, gameViews, spriteShip));
         shipCollisionDetector.start();
     }
-
     public boolean onTouch(View view, MotionEvent event) {
         switch (view.getId()) {
             case R.id.button_der:
@@ -156,21 +156,12 @@ public class GameActivity extends PlayActivity implements View.OnTouchListener {
             collisionDetectorThread.start();
         }
     }
-    public void cambiarColor(){
-        Random rnd = new Random();
-        int color = Color.argb(255, rnd.nextInt(256), rnd.nextInt(256), rnd.nextInt(256));
-        for (ImageView m:this.marcianitos.getVistasMarcianos()) {
-            m.setColorFilter(color);
-        }
-    }
+
     public void kill(final Object collider1, final ImageView collider2) {
         this.runOnUiThread(new Runnable() {
             @Override
             public void run() {
                 if (collider1 instanceof Bullet) {
-                    if (((BitmapDrawable) collider2.getDrawable()).getBitmap().sameAs(((BitmapDrawable) getResources().getDrawable(R.drawable.defensas)).getBitmap())){
-                        cambiarColor();
-                    }
                     ((Bullet) collider1).delete();
                 }
                 if (collider1 instanceof Defensas) {
